+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/AddressElement.kt
@@ -2,11 +2,10 @@ package com.stripe.android.ui.core.elements
 
 import androidx.annotation.RestrictTo
 import androidx.annotation.VisibleForTesting
-import androidx.compose.ui.text.intl.Locale
-import androidx.compose.ui.text.toLowerCase
 import com.stripe.android.ui.core.R
 import com.stripe.android.ui.core.address.AddressRepository
 import com.stripe.android.ui.core.elements.autocomplete.DefaultIsPlacesAvailable
+import com.stripe.android.ui.core.elements.autocomplete.IsPlacesAvailable
 import com.stripe.android.uicore.elements.CountryConfig
 import com.stripe.android.uicore.elements.DropdownFieldController
 import com.stripe.android.uicore.elements.PhoneNumberController
@@ -20,7 +19,6 @@ import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flattenConcat
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onEach
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 open class AddressElement constructor(
@@ -34,7 +32,8 @@ open class AddressElement constructor(
         rawValuesMap[IdentifierSpec.Country]
     ),
     sameAsShippingElement: SameAsShippingElement?,
-    shippingValuesMap: Map<IdentifierSpec, String?>?
+    shippingValuesMap: Map<IdentifierSpec, String?>?,
+    private val isPlacesAvailable: IsPlacesAvailable = DefaultIsPlacesAvailable(),
 ) : SectionMultiFieldElement(_identifier) {
 
     @VisibleForTesting
@@ -75,11 +74,13 @@ open class AddressElement constructor(
             countryCode?.let {
                 phoneNumberElement.controller.countryDropdownController.onRawValueChange(it)
             }
-            addressRepository.get(countryCode)
-                ?: emptyList()
-        }
-        .onEach { fields ->
-            fields.forEach { field ->
+            (addressRepository.get(countryCode) ?: emptyList()).onEach { field ->
+                updateLine1WithAutocompleteAffordance(
+                    field = field,
+                    countryCode = countryCode,
+                    addressType = addressType,
+                    isPlacesAvailable = isPlacesAvailable,
+                )
                 field.setRawValue(rawValuesMap)
             }
         }
@@ -166,13 +167,7 @@ open class AddressElement constructor(
             is AddressType.ShippingCondensed -> {
                 // If the merchant has supplied Google Places API key, Google Places SDK is
                 // available, and country is supported, use autocomplete
-                val supportedCountries = addressType.autocompleteCountries
-                val autocompleteSupportsCountry = supportedCountries
-                    ?.map { it.toLowerCase(Locale.current) }
-                    ?.contains(country?.toLowerCase(Locale.current)) == true
-                val autocompleteAvailable = DefaultIsPlacesAvailable().invoke() &&
-                    !addressType.googleApiKey.isNullOrBlank()
-                if (autocompleteSupportsCountry && autocompleteAvailable) {
+                if (addressType.supportsAutoComplete(country, isPlacesAvailable)) {
                     condensed
                 } else {
                     expanded
@@ -229,4 +224,48 @@ open class AddressElement constructor(
     override fun setRawValue(rawValuesMap: Map<IdentifierSpec, String?>) {
         this.rawValuesMap = rawValuesMap
     }
+}
+
+internal suspend fun updateLine1WithAutocompleteAffordance(
+    field: SectionFieldElement,
+    countryCode: String?,
+    addressType: AddressType,
+    isPlacesAvailable: IsPlacesAvailable,
+) {
+    if (field.identifier == IdentifierSpec.Line1) {
+        val fieldController = (field as? SimpleTextElement)?.controller
+        val config = (fieldController as? SimpleTextFieldController?)?.textFieldConfig
+        val textConfig = config as? SimpleTextFieldConfig?
+        if (textConfig != null) {
+            updateLine1ConfigForAutocompleteAffordance(
+                textConfig = textConfig,
+                countryCode = countryCode,
+                addressType = addressType,
+                isPlacesAvailable = isPlacesAvailable,
+            )
+        }
+    }
+}
+
+private suspend fun updateLine1ConfigForAutocompleteAffordance(
+    textConfig: SimpleTextFieldConfig,
+    countryCode: String?,
+    addressType: AddressType,
+    isPlacesAvailable: IsPlacesAvailable,
+) {
+    val supportsAutocomplete = (addressType as? AutocompleteCapableAddressType)
+        ?.supportsAutoComplete(countryCode, isPlacesAvailable)
+    val icon: TextFieldIcon.Trailing? = if (supportsAutocomplete == true) {
+        TextFieldIcon.Trailing(
+            idRes = R.drawable.stripe_ic_search,
+            isTintable = true,
+            contentDescription = R.string.address_search_content_description,
+            onClick = {
+                addressType.onNavigation()
+            }
+        )
+    } else {
+        null
+    }
+    textConfig.trailingIcon.emit(icon)
 }