+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentSheetViewModelInjectionTest.kt
@@ -12,15 +12,11 @@ import com.stripe.android.ui.core.address.AddressRepository
 import com.stripe.android.ui.core.forms.resources.LpmRepository
 import com.stripe.android.utils.PaymentIntentFactory
 import com.stripe.android.utils.fakeCreationExtras
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.FlowPreview
 import org.junit.After
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 
-@OptIn(FlowPreview::class)
-@ExperimentalCoroutinesApi
 @RunWith(RobolectricTestRunner::class)
 internal class PaymentSheetViewModelInjectionTest : BasePaymentSheetViewModelInjectionTest() {