+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/addresselement/AddressElementViewModelTest.kt
@@ -6,7 +6,6 @@ import com.google.common.truth.Truth
 import com.stripe.android.core.injection.Injectable
 import com.stripe.android.core.injection.NonFallbackInjector
 import com.stripe.android.core.injection.WeakMapInjectorRegistry
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -19,7 +18,6 @@ import org.mockito.kotlin.verify
 import org.robolectric.RobolectricTestRunner
 import kotlin.test.assertNotNull
 
-@ExperimentalCoroutinesApi
 @RunWith(RobolectricTestRunner::class)
 class AddressElementViewModelTest {
     private val defaultArgs = AddressElementActivityContract.Args(